==> code/1. Yahoo_Finance_Crawl.ipynb <==
import yfinance as yf
import pandas as pd

#2021/02/04 NASDAQ, NYSE SYMBOL Load
NASDAQ_sb = pd.read_csv("../data/NASDAQ_MC10B_210203.csv")

#Sort by Market Cap 100
NASDAQ100 = NASDAQ_sb.sort_values("Market Cap", ascending=False)[:100]
NASDAQ100

NASDAQ100[(NASDAQ100["IPO Year"]<2009) | (NASDAQ100["IPO Year"].isna())]

#Load NASDAQ100 2009~2021
NASDAQ100_Price = yf.download(NASDAQ100.Symbol.tolist(), start="2009-01-02")

NASDAQ100_Price = NASDAQ100_Price.dropna(axis=1)

def tuple_join(tuple_var):
    return(tuple_var[1]+" "+tuple_var[0])

def multicol_join_sort(multicol_df):
    multicol_df.columns = list(map(tuple_join, multicol_df.columns))
    multicol_df = multicol_df[sorted(multicol_df.columns.tolist())]
    multicol_df.reset_index(inplace=True)
    return(multicol_df)

NASDAQ100_Price = multicol_join_sort(NASDAQ100_Price)
NASDAQ100_Price

NASDAQ100_Price.to_csv("../data/NASDAQ100.csv", index=False)

NASDAQ_2009 = list(set(list(map(lambda x: x.split()[0],NASDAQ100_Price.columns.tolist()))))

NASDAQ100_2009 = NASDAQ100[NASDAQ100["Symbol"].isin(NASDAQ_2009)]
NASDAQ100_2009

NASDAQ100_2009.to_csv("../data/NASDAQ100_2009_CUT.csv", index=False)
