==> code/2. NASDAQ_TA-Lib_Extraction.ipynb <==
from talib import *
import pandas as pd

NASDAQ_stock = pd.read_csv("../data/NASDAQ100.csv")
NASDAQ_symbol = pd.read_csv("../data/NASDAQ100_2009_CUT.csv")["Symbol"].to_list()

def TI_ex(symbol, OHLCV_data):
    Date = OHLCV_data["Date"]
    Adj_Close = OHLCV_data[symbol+" Adj Close"]
    Close = OHLCV_data[symbol+" Close"]
    High = OHLCV_data[symbol+" High"]
    Low = OHLCV_data[symbol+" Low"]
    Open = OHLCV_data[symbol+" Open"]
    Volume = OHLCV_data[symbol+" Volume"]
    
    adx = ADX(High, Low, Close)
    aroondown, aroonup = AROON(High, Low)
    aroonosc = AROONOSC(High, Low)
    bop = BOP(Open, High, Low, Close)
    cci = CCI(High, Low, Close, timeperiod=14)
    cmo = CMO(Adj_Close, timeperiod=14)
    dx = DX(High, Low, Close)
    mfi = MFI(High, Low, Close, Volume)
    minus_di = MINUS_DI(High, Low, Close)
    plus_di = PLUS_DI(High, Low, Close)
    ppo = PPO(Adj_Close)
    roc = ROC(Adj_Close, timeperiod=12)
    rsi = RSI(Adj_Close)
    slowk, slowd = STOCH(High, Low, Close, fastk_period=12)
    fastk, fastd = STOCHF(High, Low, Close, fastk_period=12)
    trix = TRIX(Adj_Close, timeperiod=12)
    ultosc = ULTOSC(High, Low, Close)
    willr = WILLR(High, Low, Close)
    
    TI_DF = pd.DataFrame({
        "Date":Date,
        "Adj_Close":Adj_Close,
        "Open":Open,
        "adx":adx,
        "aroondown":aroondown,
        "aroonup":aroonup,
        "aroonosc":aroonosc,
        "bop":bop,
        "cci":cci,
        "cmo":cmo,
        "dx":dx,
        "mfi":mfi,
        "minus_di":minus_di,
        "plus_di":plus_di,
        "ppo":ppo,
        "roc":roc,
        "rsi":rsi,
        "slowk":slowk,
        "slowd":slowd,
        "fastk":fastk,
        "fastd":fastd,
        "trix":trix,
        "ultosc":ultosc,
        "willr":willr
    })
    
    TI_DF.to_csv("../data/Stock_TI/"+symbol+"_TI.csv", index=False)
    print(symbol+" Technical Indicator Extraction!\nshape =",TI_DF.shape)

for symbol in NASDAQ_symbol: TI_ex(symbol, NASDAQ_stock)
